fix: item 가격 데이터타입 변경 (Long -> int)

+++ dolharubang/src/main/java/com/dolharubang/mongo/entity/Item.java
@@ -22,7 +22,7 @@ public class Item {
 
     private String imageUrl;
 
-    private Long price;
+    private int price;
 
     @Field
     private LocalDateTime createdAt;
@@ -30,7 +30,7 @@ public class Item {
     @Field
     private LocalDateTime modifiedAt;
 
-    public Item update(ItemType itemType, String itemName, String imageUrl, Long price) {
+    public Item update(ItemType itemType, String itemName, String imageUrl, int price) {
         return Item.builder()
             .itemId(this.itemId)
             .itemType(itemType)
